requirements.txt
aiofiles==23.2.1
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cycler==0.12.1
etelemetry==0.3.1
exceptiongroup==1.2.1
filelock==3.13.4
fitz==0.0.1.dev2
fonttools==4.51.0
frontend==0.0.3
h11==0.14.0
httplib2==0.22.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
kiwisolver==1.4.5
llvmlite==0.42.0
looseversion==1.3.0
lxml==5.2.1
matplotlib==3.9.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numba==0.59.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pillow==10.3.0
prov==2.0.0
pydot==2.0.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pyxnat==1.6.2
rdflib==7.0.0
requests==2.31.0
scipy==1.13.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
traits==6.3.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wget==3.2